+++ cobalt/browser/web_module.cc
@@ -799,8 +799,8 @@ void WebModule::Impl::InjectCustomWindowAttributes(
            attributes.begin();
        iter != attributes.end(); ++iter) {
     global_environment_->Bind(
-        iter->first,
-        iter->second.Run(window_, &mutation_observer_task_manager_));
+        iter->first, iter->second.Run(window_, &mutation_observer_task_manager_,
+                                      global_environment_.get()));
   }
 }